+++ lib/mongoid/copyable.rb
@@ -5,15 +5,6 @@ module Mongoid
   module Copyable
     extend ActiveSupport::Concern
 
-    COPYABLES = [
-      :attributes,
-      :metadata,
-      :changed_attributes,
-      :previous_changes
-    ]
-
-    protected
-
     # Clone or dup the current +Document+. This will return all attributes with
     # the exception of the document's id and versions, and will reset all the
     # instance variables.
@@ -26,53 +17,13 @@ module Mongoid
     # @param [ Document ] other The document getting cloned.
     #
     # @return [ Document ] The new document.
-    def initialize_copy(other)
-      __copy__(other)
-    end
-
-    # Clone or dup the current +Document+. This will return all attributes with
-    # the exception of the document's id and versions, and will reset all the
-    # instance variables.
-    #
-    # This clone also includes embedded documents.
-    #
-    # @example Dup the document.
-    #   document.dup
-    #
-    # @param [ Document ] other The document getting cloned.
-    #
-    # @return [ Document ] The new document.
-    def initialize_dup(other)
-      __copy__(other)
-    end
-
-    private
-
-    # Handle the copy of the object.
-    #
-    # @api private
-    #
-    # @example Copy the object.
-    #   document.__copy__(other)
-    #
-    # @param [ Document ] other The other document.
-    #
-    # @return [ Document ] self.
-    #
-    # @since 3.0.0
-    def __copy__(other)
-      other.as_document
-      instance_variables.each { |name| remove_instance_variable(name) }
-      COPYABLES.each do |name|
-        value = other.send(name)
-        instance_variable_set(:"@#{name}", value ? value.dup : nil)
+    def clone
+      attrs = as_document.except("_id")
+      if attrs.delete("versions")
+        attrs["version"] = 1
       end
-      attributes.delete("_id")
-      if attributes.delete("versions")
-        attributes["version"] = 1
-      end
-      @new_record = true
-      apply_defaults
+      self.class.new(attrs)
     end
+    alias :dup :clone
   end
 end